+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v12.0.4
+* Resource Display now have secret-compatible percentage threshold settings for applying custom Opacity & Color above or below certain values.
+
 ## v12.0.3
 * Groups can now inherit their opacity from an icon. This allows an entire group to take on opacity that is driven by a secret aspect, like cooldown readiness or spell casting state. This also allows using an icon's conditions to drive group opacity.
 * The Spell Cast condition can detect secret casts if the spell filter is empty and the "Casting" checkbox is chosen.